Added service status updaters

+++ module_search/service_search.py
@@ -2,6 +2,9 @@ import urllib.request
 import os
 import requests
 
+import module_search.callback_agent
+
+
 def download(url: str, dest_folder: str):
     if not os.path.exists(dest_folder):
         os.makedirs(dest_folder)  # create folder if it does not exist
@@ -22,24 +25,35 @@ def download(url: str, dest_folder: str):
         return file_path
     else:  # HTTP status code 4XX/5XX
         print("Download failed: status code {}\n{}".format(r.status_code, r.text))
+        return ""
 
 def unzip(directory: str):
     print("Unzipping in dir: ", directory)
+
+    if not os.path.exists(directory):
+        return False
+
     os.system("apt-get install unzip")
 
     print(os.system("unzip -o "+directory))
 
     print("Execution complete!")
 
+    return True
+
 def execute_playbook_script(directory: str):
     print("executing Playbook in dir: ", directory)
 
-    print(os.system("ansible-playbook "+directory))
+    if not os.path.exists(directory):
+        return False
 
+    print(os.system("ansible-playbook -i hosts "+directory)) # haven't tried with -i hosts flag
     print("Execution complete!")
 
+    return True
+
 class plusclouds_service:
-    def __init__(self, service_name: str, service_url: str):
+    def __init__(self, service_name: str, service_url: str, callback_ansible_url: str, callback_service_url: str):
         self.service_name = service_name
 
         self.download_path = "./"+service_name
@@ -50,6 +64,11 @@ class plusclouds_service:
 
         self.service_url = service_url
 
+        self.callback_ansible_url = callback_ansible_url
+
+        self.callback_service_url = callback_service_url
+
+        self.callback_agent = module_search.callback_agent.CallbackAgent(self.callback_service_url)
 
     def download_module(self):
         if self.is_downloaded:
@@ -58,13 +77,32 @@ class plusclouds_service:
         self.is_downloaded = True
         return download(self.service_url, self.download_path)
 
+
     def initiate_ansible(self):
         if not self.is_downloaded:
             self.download_module()
+
         execute_playbook_script("install.yml")
         self.is_initiated = True
 
+
     def run(self):
-        path = self.download_module()
-        unzip(path)
+        self.callback_agent.downloading("Downloading starting")
+        try:
+            path = self.download_module()
+        except requests.exceptions.ConnectionError as e:
+            self.callback_agent.failed("Download failed ")
+            return
+
+        else:
+            self.callback_agent.initiating("Download completed, starting unzipping")
+
+        if not unzip(path):
+            self.callback_agent.failed("Unzip failed")
+            return
+        else:
+            self.callback_agent.initiating("Playbook Execution starting.")
+
+
         self.initiate_ansible()
+        self.callback_agent.completed("Completed!")
